+++ cohort_creator/_cli.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from typing import Sequence
 
 import pandas as pd
+from datalad import api
 
 from cohort_creator._parsers import global_parser
 from cohort_creator._utils import get_bids_filter
@@ -46,6 +47,13 @@ def _get_participant_listing_from_args(args: argparse.Namespace) -> pd.DataFrame
     return load_participant_listing(participant_listing=participant_listing)
 
 
+def create_yoda(output_dir: Path) -> None:
+    if not output_dir.exists():
+        api.create(path=output_dir, cfg_proc="yoda")
+    if not (output_dir / ".datalad").exists():
+        api.create(path=output_dir, cfg_proc="yoda", force=True)
+
+
 def cli(argv: Sequence[str] = sys.argv) -> None:
     """Entry point."""
     parser = global_parser()
@@ -64,11 +72,10 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
 
     dataset_listing = load_dataset_listing(dataset_listing=args.dataset_listing)
 
-    sourcedata_dir = output_dir / "sourcedata"
-    sourcedata_dir.mkdir(exist_ok=True, parents=True)
-
     if args.command in ["install", "all"]:
-        _execute_install(dataset_listing, args, sourcedata_dir)
+        create_yoda(output_dir)
+        (output_dir / "sourcedata").mkdir(exist_ok=True, parents=True)
+        _execute_install(dataset_listing, args, output_dir)
     if args.command == "install":
         return None
 
@@ -82,7 +89,7 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
         if isinstance(jobs, list):
             jobs = jobs[0]
         get_data(
-            sourcedata=sourcedata_dir,
+            output_dir=output_dir,
             datasets=dataset_listing,
             participants=participant_listing,
             dataset_types=dataset_types,
@@ -98,7 +105,6 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
         skip_group_mriqc = bool(args.skip_group_mriqc)
         construct_cohort(
             output_dir=output_dir,
-            sourcedata_dir=sourcedata_dir,
             datasets=dataset_listing,
             participants=participant_listing,
             dataset_types=dataset_types,
@@ -111,18 +117,22 @@ def cli(argv: Sequence[str] = sys.argv) -> None:
 
 
 def _execute_install(
-    dataset_listing: pd.DataFrame, args: argparse.Namespace, sourcedata_dir: Path
+    dataset_listing: pd.DataFrame, args: argparse.Namespace, output_dir: Path
 ) -> None:
     participant_listing = _get_participant_listing_from_args(args)
+
     datasets_to_install = get_list_datasets_to_install(
         dataset_listing=dataset_listing, participant_listing=participant_listing
     )
+
     generate_participant_listing = getattr(args, "generate_participant_listing", False)
+
     if participant_listing is None:
         generate_participant_listing = True
+
     install_datasets(
         datasets=datasets_to_install,
-        sourcedata=sourcedata_dir,
+        output_dir=output_dir,
         dataset_types=args.dataset_types,
         generate_participant_listing=generate_participant_listing,
     )