+++ tests/test_cohort_creator.py
@@ -2,41 +2,48 @@
 from __future__ import annotations
 
 import pandas as pd
+import pytest
 
+from cohort_creator._cli import create_yoda
+from cohort_creator._utils import sourcedata
 from cohort_creator.cohort_creator import construct_cohort
 from cohort_creator.cohort_creator import get_data
 from cohort_creator.cohort_creator import install_datasets
 
 
-def test_install_datasets(tmp_path, caplog):
-    sourcedata = tmp_path / "sourcedata"
+@pytest.fixture
+def output_dir(tmp_path):
+    create_yoda(output_dir=tmp_path)
+    return tmp_path
+
+
+def test_install_datasets(output_dir, caplog):
     install_datasets(
         datasets=["ds000001", "foo"],
-        sourcedata=sourcedata,
+        output_dir=output_dir,
         dataset_types=["raw", "mriqc", "fmriprep"],
     )
     assert "foo not found in openneuro" in caplog.text
-    assert (sourcedata / "ds000001").exists()
-    assert (sourcedata / "ds000001-mriqc").exists()
-    assert (sourcedata / "ds000001-fmriprep").exists()
+    assert (sourcedata(output_dir) / "ds000001").exists()
+    assert (sourcedata(output_dir) / "ds000001-mriqc").exists()
+    assert (sourcedata(output_dir) / "ds000001-fmriprep").exists()
 
     # caplog.clear()
     # install_datasets(datasets=["ds000001"], sourcedata=sourcedata, dataset_types= ["raw"])
     # assert "data already present at" in caplog.text
 
 
-def test_install_datasets_create_participant_listing(tmp_path):
-    sourcedata = tmp_path / "sourcedata"
+def test_install_datasets_create_participant_listing(output_dir):
     install_datasets(
         datasets=["ds000002"],
-        sourcedata=sourcedata,
+        output_dir=output_dir,
         dataset_types=["raw"],
         generate_participant_listing=True,
     )
-    (tmp_path / "code" / "participants.tsv").exists()
+    (output_dir / "code" / "participants.tsv").exists()
 
 
-def test_construct_cohort(tmp_path):
+def test_construct_cohort(output_dir):
     participants = pd.DataFrame(
         {"DatasetID": ["ds000001"], "SubjectID": ["sub-01"], "SessionID": [""]}
     )
@@ -46,15 +53,13 @@ def test_construct_cohort(tmp_path):
             "PortalURI": ["https://github.com/OpenNeuroDatasets-JSONLD/ds000001.git"],
         }
     )
-    output_dir = tmp_path / "outputs"
-    sourcedata = output_dir / "sourcedata"
     dataset_types = ["raw"]
     datatypes = ["anat"]
     install_datasets(
-        datasets=["ds000001", "foo"], sourcedata=sourcedata, dataset_types=dataset_types
+        datasets=["ds000001", "foo"], output_dir=output_dir, dataset_types=dataset_types
     )
     get_data(
-        sourcedata=sourcedata,
+        output_dir=output_dir,
         datasets=datasets,
         participants=participants,
         dataset_types=dataset_types,
@@ -63,7 +68,7 @@ def test_construct_cohort(tmp_path):
         jobs=2,
     )
     get_data(
-        sourcedata=sourcedata,
+        output_dir=output_dir,
         datasets=datasets,
         participants=participants,
         dataset_types=dataset_types,
@@ -73,7 +78,6 @@ def test_construct_cohort(tmp_path):
     )
     construct_cohort(
         output_dir=output_dir,
-        sourcedata_dir=sourcedata,
         datasets=datasets,
         participants=participants,
         dataset_types=dataset_types,
@@ -82,7 +86,6 @@ def test_construct_cohort(tmp_path):
     )
     construct_cohort(
         output_dir=output_dir,
-        sourcedata_dir=sourcedata,
         datasets=datasets,
         participants=participants,
         dataset_types=dataset_types,